SignedDistanceFunctions

Documentation for SignedDistanceFunctions.

SignedDistanceFunctions.CentredGeometry

SignedDistanceFunctions.Cube

SignedDistanceFunctions.Geometry

SignedDistanceFunctions.Sphere

Base.intersect

Base.union

SignedDistanceFunctions.scale

SignedDistanceFunctions.signed_distance

SignedDistanceFunctions.CentredGeometry — Type

CentredGeometry{D} <: Geometry{D}

Geometry that is centred on the origin.

SignedDistanceFunctions.Cube — Type

Cube{D}

The unit cube centred on the origin in D dimensions.

SignedDistanceFunctions.Geometry — Type

Geometry{D}

Abstract type representing a D-dimensional geometry.

SignedDistanceFunctions.Sphere — Type

Sphere{D}

The unit sphere centred on the origin in D dimensions.

Base.intersect — Method

intersect(g::Geometry, rest...)

Create a geometry which is the intersection of all given geometries.

Base.union — Method

union(g::Geometry, rest...)

Create a geometry which is the union of all given geometries.

SignedDistanceFunctions.scale — Method

scale(g::Geometry, factor::Float64)
scale(g::Geometry, factor::NTuple)

Scale the given geometry by factor.

If the factor is a scalar, this will be a

SignedDistanceFunctions.signed_distance — Function

signed_distance(geometry, x)

Get the signed distance from position x to geometry.

Args

geometry::Geometry{D}: An object representing the geometry.

x::NTuple{Real,D}: The position in space.